fix(topics): reliable per-ticker institutional flow + rounded numbers
Surfaced while verifying against real FinMind data:

- The 三大法人 flow was all-zeros on a cold path: the single full-market
  bulk request gets rate-limited (returns empty) when it runs after the
  other FinMind fetches. Switch to per-ticker fetches (same proven pattern
  as trading-value), fetch institutional FIRST while the budget is fresh,
  and never cache an empty result (so a transient miss can't poison the
  day-long cache). Real magnitudes now calibrate correctly vs TWSE T86
  (先進封裝 5d 外資 +216億; AI 伺服器 -363億).
- Round all performance numbers server-side: heat 1dp, return 2dp, and NT$
  amounts as plain ints (schema typed int so Pydantic stops coercing the
  rounded values back to .0 floats). Clean numbers reach the chart/tooltips.

Verified: py_compile + institutional aggregation self-test (updated for the
per-ticker mock) green.

+++ backend/src/schemas/sector.py
@@ -78,14 +78,15 @@ class ExposurePerformanceItem(BaseModel):
     display_name: str
     color_hex: Optional[str] = None
     episode_count: int = 0
-    heat: Optional[float] = None               # recency-weighted discussion (X axis)
-    return_pct: Optional[float] = None         # avg member daily % change
-    market_cap_twd: Optional[float] = None     # aggregate constituent market cap (industries only)
-    trading_value_twd: Optional[float] = None  # aggregate constituent daily trade value (NT$)
-    trading_value_windows_twd: Optional[Dict[str, float]] = None
+    heat: Optional[float] = None               # recency-weighted discussion (X axis), 1 dp
+    return_pct: Optional[float] = None         # avg member daily % change, 2 dp
+    # NT$ amounts are whole numbers (rounded server-side) — int keeps the JSON clean (no .0).
+    market_cap_twd: Optional[int] = None       # aggregate constituent market cap (industries only)
+    trading_value_twd: Optional[int] = None    # aggregate constituent daily trade value (NT$)
+    trading_value_windows_twd: Optional[Dict[str, int]] = None
     # 三大法人 net flow by window (1/5/20d), NT$ — TW members only. Total = all法人; foreign = 外資.
-    net_buy_windows_twd: Optional[Dict[str, float]] = None
-    foreign_net_windows_twd: Optional[Dict[str, float]] = None
+    net_buy_windows_twd: Optional[Dict[str, int]] = None
+    foreign_net_windows_twd: Optional[Dict[str, int]] = None
 
 
 class ExposurePerformanceResponse(BaseModel):
